Move cgo merkle tree to alternative package.
This allows 3rd parties to use merkle tree verification algorithms
without aquiring a dependency on openssl.
