+++ go/fulltree/merkle_tree_go.cc
