+++ go/fulltree/merkle_tree_interface.go
@@ -1,4 +1,4 @@
-package merkletree
+package fulltree
 
 // MerkleTreeInterface represents the common interface for basic MerkleTree functions.
 type MerkleTreeInterface interface {
