+++ go/fulltree/merkle_tree_test.go
@@ -1,4 +1,4 @@
-package merkletree
+package fulltree
 
 import (
 	"bytes"
