+++ go/fulltree/wrap_merkle_tree.cc
