+++ go/fulltree/wrap_merkle_tree_math.cc
