+++ go/fulltree/wrap_serial_hasher.cc
