+++ go/fulltree/wrap_tree_hasher.cc
